+++ CHANGELOG.rst
@@ -2,6 +2,8 @@ Changelog
 =========
 
 ## 0.18.1
+  * Densities are now loaded using float64 precision. Higher precision aids in calculating
+    correctly the total counts when accumulation of small and large densities is involved.
   * `layer` made an optional trait in `brainbuilder.app.cells.place`
   * `subregion` added as cells property in `brainbuilder.app.cells.place`